+++ src/main/java/real/world/domain/user/service/UserService.java
@@ -1,13 +1,9 @@
 package real.world.domain.user.service;
 
-import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
 import real.world.domain.user.dto.request.RegisterRequest;
 import real.world.domain.user.dto.response.LoginResponse;
 import real.world.domain.user.dto.response.RegisterResponse;
@@ -15,7 +11,6 @@ import real.world.domain.user.entity.User;
 import real.world.domain.user.repository.UserRepository;
 import real.world.error.exception.JwtInvalidException;
 import real.world.error.exception.UsernameAlreadyExistsException;
-import real.world.security.JwtUtil;
 
 @Service
 public class UserService {
