+++ src/main/java/real/world/security/UserDetailsByEmailService.java
@@ -3,23 +3,24 @@ package real.world.security;
 import java.util.Collections;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 import real.world.domain.user.entity.User;
 import real.world.domain.user.repository.UserRepository;
 import real.world.error.ErrorCode;
 import real.world.error.exception.AuthenticationErrorCodeException;
+import real.world.security.authentication.CustomUserDetails;
 
 @RequiredArgsConstructor
 @Service
-public class UserDetailsByEmailService implements UserDetailsService {
+public class UserDetailsByEmailService implements CustomUserDetailsService {
 
     private final UserRepository userRepository;
 
     @Override
-    public UserDetails loadUserByUsername(String email) {
+    public UserDetails loadUserByPrincipal(String email) throws AuthenticationException {
         User user = userRepository.findByEmail(email)
             .orElseThrow(
                 () -> new AuthenticationErrorCodeException(ErrorCode.USERNAME_NOT_EXIST.toString(), ErrorCode.USERNAME_NOT_EXIST));
