+++ src/main/java/real/world/security/authentication/CustomAuthenticationProvider.java
@@ -1,18 +1,18 @@
-package real.world.security;
+package real.world.security.authentication;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import real.world.error.ErrorCode;
 import real.world.error.exception.AuthenticationErrorCodeException;
+import real.world.security.CustomUserDetailsService;
 
 @RequiredArgsConstructor
 public class CustomAuthenticationProvider implements AuthenticationProvider {
 
-    private final UserDetailsService userDetailsService;
+    private final CustomUserDetailsService userDetailsService;
 
     private final PasswordEncoder passwordEncoder;
 
@@ -23,7 +23,7 @@ public class CustomAuthenticationProvider implements AuthenticationProvider {
         final String email = token.getName();
         final String password = (String) token.getCredentials();
 
-        final CustomUserDetails userDetails = (CustomUserDetails) userDetailsService.loadUserByUsername(email);
+        final CustomUserDetails userDetails = (CustomUserDetails) userDetailsService.loadUserByPrincipal(email);
         if (!passwordEncoder.matches(password, userDetails.getPassword())) {
             throw new AuthenticationErrorCodeException(ErrorCode.WRONG_PASSWORD.toString(), ErrorCode.WRONG_PASSWORD);
         }
