+++ src/main/java/real/world/security/jwt/JwtUtil.java
@@ -1,4 +1,4 @@
-package real.world.security;
+package real.world.security.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
